ai-spacemaker/llm-cohort2/src/2_rag_pure_python/requirements.txt
chainlit==0.7.0
openai==0.28.0
python-dotenv==1.0.0
wandb
numpy
pandas
matplotlib
plotly
scipy
scikit-learn